requirements.txt
aiomysql
sshtunnel
discord
pillow
aiofiles
colorlog
numpy
scipy
apscheduler
fuzzywuzzy[speedup]
